closes #5

+++ backend/ADR_Viewer/src/main/java/com/serviceEngineering/ADR_Viewer/repository/ADRRepository.java
@@ -3,5 +3,5 @@ package com.serviceEngineering.ADR_Viewer.repository;
 import com.serviceEngineering.ADR_Viewer.entity.ADR;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ADRRepository extends JpaRepository<ADR, Integer> {
+public interface ADRRepository extends JpaRepository<ADR, Long> {
 }
